Add additional sync entity tests
